+++ tests/SurfaceFactoryTest.cpp
@@ -1,15 +1,15 @@
-// Copyright 2017-2019 the Polygon Mesh Processing Library developers.
+// Copyright 2017-2021 the Polygon Mesh Processing Library developers.
 // Distributed under a MIT-style license, see LICENSE.txt for details.
 
 #include "gtest/gtest.h"
 
-#include <pmp/algorithms/SurfacePrimitives.h>
+#include <pmp/algorithms/SurfaceFactory.h>
 
 using namespace pmp;
 
 bool vertices_on_sphere(const SurfaceMesh& mesh)
 {
-    for (auto v: mesh.vertices())
+    for (auto v : mesh.vertices())
     {
         const Scalar n = norm(mesh.position(v));
         if (n < 0.999 || n > 1.001)
@@ -20,9 +20,9 @@ bool vertices_on_sphere(const SurfaceMesh& mesh)
     return true;
 }
 
-TEST(SurfacePrimitivesTest, tetrahedron)
+TEST(SurfaceFactoryTest, tetrahedron)
 {
-    auto mesh = tetrahedron();
+    auto mesh = SurfaceFactory::tetrahedron();
     EXPECT_EQ(mesh.n_vertices(), 4u);
     EXPECT_EQ(mesh.n_edges(), 6u);
     EXPECT_EQ(mesh.n_faces(), 4u);
@@ -30,9 +30,9 @@ TEST(SurfacePrimitivesTest, tetrahedron)
     EXPECT_TRUE(vertices_on_sphere(mesh));
 }
 
-TEST(SurfacePrimitivesTest, octahedron)
+TEST(SurfaceFactoryTest, octahedron)
 {
-    auto mesh = octahedron();
+    auto mesh = SurfaceFactory::octahedron();
     EXPECT_EQ(mesh.n_vertices(), 6u);
     EXPECT_EQ(mesh.n_edges(), 12u);
     EXPECT_EQ(mesh.n_faces(), 8u);
@@ -40,9 +40,9 @@ TEST(SurfacePrimitivesTest, octahedron)
     EXPECT_TRUE(vertices_on_sphere(mesh));
 }
 
-TEST(SurfacePrimitivesTest, hexahedron)
+TEST(SurfaceFactoryTest, hexahedron)
 {
-    auto mesh = hexahedron();
+    auto mesh = SurfaceFactory::hexahedron();
     EXPECT_EQ(mesh.n_vertices(), 8u);
     EXPECT_EQ(mesh.n_edges(), 12u);
     EXPECT_EQ(mesh.n_faces(), 6u);
@@ -50,9 +50,9 @@ TEST(SurfacePrimitivesTest, hexahedron)
     EXPECT_TRUE(vertices_on_sphere(mesh));
 }
 
-TEST(SurfacePrimitivesTest, icosahedron)
+TEST(SurfaceFactoryTest, icosahedron)
 {
-    auto mesh = icosahedron();
+    auto mesh = SurfaceFactory::icosahedron();
     EXPECT_EQ(mesh.n_vertices(), 12u);
     EXPECT_EQ(mesh.n_edges(), 30u);
     EXPECT_EQ(mesh.n_faces(), 20u);
@@ -60,9 +60,9 @@ TEST(SurfacePrimitivesTest, icosahedron)
     EXPECT_TRUE(vertices_on_sphere(mesh));
 }
 
-TEST(SurfacePrimitivesTest, dodecahedron)
+TEST(SurfaceFactoryTest, dodecahedron)
 {
-    auto mesh = dodecahedron();
+    auto mesh = SurfaceFactory::dodecahedron();
     EXPECT_EQ(mesh.n_vertices(), 20u);
     EXPECT_EQ(mesh.n_edges(), 30u);
     EXPECT_EQ(mesh.n_faces(), 12u);
@@ -71,9 +71,9 @@ TEST(SurfacePrimitivesTest, dodecahedron)
     EXPECT_TRUE(vertices_on_sphere(mesh));
 }
 
-TEST(SurfacePrimitivesTest, icosphere)
+TEST(SurfaceFactoryTest, icosphere)
 {
-    auto mesh = icosphere();
+    auto mesh = SurfaceFactory::icosphere();
     EXPECT_EQ(mesh.n_vertices(), 642u);
     EXPECT_EQ(mesh.n_edges(), 1920u);
     EXPECT_EQ(mesh.n_faces(), 1280u);
@@ -81,9 +81,9 @@ TEST(SurfacePrimitivesTest, icosphere)
     EXPECT_TRUE(vertices_on_sphere(mesh));
 }
 
-TEST(SurfacePrimitivesTest, quad_sphere)
+TEST(SurfaceFactoryTest, quad_sphere)
 {
-    auto mesh = quad_sphere();
+    auto mesh = SurfaceFactory::quad_sphere();
     EXPECT_EQ(mesh.n_vertices(), 386u);
     EXPECT_EQ(mesh.n_edges(), 768u);
     EXPECT_EQ(mesh.n_faces(), 384u);
@@ -91,9 +91,9 @@ TEST(SurfacePrimitivesTest, quad_sphere)
     EXPECT_TRUE(vertices_on_sphere(mesh));
 }
 
-TEST(SurfacePrimitivesTest, uv_sphere)
+TEST(SurfaceFactoryTest, uv_sphere)
 {
-    auto mesh = uv_sphere();
+    auto mesh = SurfaceFactory::uv_sphere();
     EXPECT_EQ(mesh.n_vertices(), 212u);
     EXPECT_EQ(mesh.n_edges(), 435u);
     EXPECT_EQ(mesh.n_faces(), 225u);